Remove math_objects and move its contents to src/fintamath
