+++ src/fintamath/MathObject.hpp
