+++ src/fintamath/nodes/expressions/Expression.hpp
@@ -3,7 +3,7 @@
 #include <memory>
 #include <vector>
 
-#include "fintamath/math_objects/nodes/Node.hpp"
+#include "fintamath/nodes/Node.hpp"
 
 namespace fintamath {
   class Expression : public Node {
