+++ src/fintamath/nodes/numbers/Rational.cpp
@@ -1,4 +1,4 @@
-#include "fintamath/math_objects/nodes/numbers/Rational.hpp"
+#include "fintamath/nodes/numbers/Rational.hpp"
 
 #include <algorithm>
 #include <stdexcept>
